+++ src/interpreter/__tests__/program.ts
@@ -1,0 +1,14 @@
+import { parse } from '../../parser/parser';
+import { compileProgram } from '../compiler';
+import { interpret } from '../virtualMachine';
+
+describe('program', () => {
+  test('handles program', () => {
+    const code = 'int main() { return 2; }';
+    const ast = parse(code);
+    const instructions = compileProgram(ast);
+    const result = interpret(instructions);
+    const expectedResult = 2;
+    expect(result).toEqual(expectedResult);
+  });
+});
