+++ src/interpreter/__tests__/variableDeclaration.ts
@@ -1,0 +1,14 @@
+import { parse } from '../../parser/parser';
+import { compileProgram } from '../compiler';
+import { interpret } from '../virtualMachine';
+
+describe('variable declaration', () => {
+  test('handles variable declarations and identifiers', () => {
+    const code = 'int main() { int a; int b = 5; return b; }';
+    const ast = parse(code);
+    const instructions = compileProgram(ast);
+    const result = interpret(instructions);
+    const expectedResult = 5;
+    expect(result).toEqual(expectedResult);
+  });
+});
