+++ src/interpreter/__tests__/whileStatement.ts
@@ -1,0 +1,40 @@
+import { parse } from '../../parser/parser';
+import { compileProgram } from '../compiler';
+import { interpret } from '../virtualMachine';
+
+describe('while statement', () => {
+  test('handles while loop', () => {
+    const code = `
+        int main() {
+            int i = 0;
+            while (i < 2) {
+                i += 1;
+            }
+            return i;
+        }
+    `;
+    const ast = parse(code);
+    const instructions = compileProgram(ast);
+    const result = interpret(instructions);
+    const expectedResult = 2;
+    expect(result).toEqual(expectedResult);
+  });
+
+  test('handles return in while loop', () => {
+    const code = `
+        int main() {
+            int i = 0;
+            while (i < 2) {
+                i += 1;
+                return i;
+            }
+            return i;
+        }
+    `;
+    const ast = parse(code);
+    const instructions = compileProgram(ast);
+    const result = interpret(instructions);
+    const expectedResult = 1;
+    expect(result).toEqual(expectedResult);
+  });
+});
